+++ src/tidyrun/__init__.py
@@ -1,24 +1,24 @@
 from importlib.metadata import version
-from .dag import (
-    DAG,
+
+from .dag import DAG, ParametrizedJob
+from .execute import (
     DAGExecutionError,
-    ParametrizedJob,
     batch_entrypoint,
     execute_plan,
     run_materialized_job,
 )
+from .executors import AwsBatchExecutor, SlurmExecutor
+from .job import Job
 from .plan import (
     PlanPaths,
     get_job_states,
     load_callable,
+    load_inputs_and_callable,
     load_job_definition,
     load_job_inputs,
     rerun_snippet,
 )
-from .executors import AwsBatchExecutor
-from .job import Job
 from .serialization import LazyDict, deserialize, serialize
-from .executors import SlurmExecutor
 
 __version__ = version("tidyrun")
 
@@ -37,6 +37,7 @@ __all__ = [
     "execute_plan",
     "get_job_states",
     "load_callable",
+    "load_inputs_and_callable",
     "load_job_definition",
     "load_job_inputs",
     "rerun_snippet",